==> 2_gwas_sumstat_preprocessing.ipynb <==
GWAS summery stat data preprocess
1.need to convert Markername to rsID by combining Chr and basepair coordinate in _lifted bim file and and replace Markername by matching
2.calculate z-score using P-value and effect size using munge_sumstat in ldsc github repository

#read meta5 file -- this is our GWAS summery stat data
! zless /data/CARD/PD/summary_stats/META5_all.gz

#copy meta 5 file into data folder - run only once for download
! cp -R /data/CARD/PD/summary_stats/META5_all.gz /data/songy4/TWAS/data/

#unzip META5_all.gz - run only once for download
! gzip -d META5_all.gz

#make sumstat.csv which contains MarkerName, Allele1, Allele2, Effect, StdErr, P-value
! awk '{print $1, $2, $3, $8, $9, $10}' data_folder/META5_all > data_folder/sumstat.txt

#import library
import subprocess
import sys
import os
import shutil
import pandas as pd
import numpy as np

#set paths
basedir = '/data/songy4/twas'
datadir = f'{basedir}/data_folder'
fusiondir = f'{basedir}/fusion_twas'
geno_path = f'{datadir}/qc_genotypes_twas'
gene_list_path = f'{datadir}/gene_list.txt'
pheno_path = f'{datadir}/expression_matrix_final.txt'
coord_path = f'{datadir}/twas_coordinate.txt'
covar_path = f'{datadir}/covariates_case.txt'

statsum = pd.read_csv(r"./data_folder/META5_all", sep='\t')
statsum.head()

#count if Freq1 is bigger than 0.5 --> this is to check if minor allele is A1
!cat {datadir}/META5_all | awk -F"\t" '$4>0.5' | wc -l

#check again how many Freq1 
statsum_flip = statsum[statsum.Freq1 > 0.5]
statsum_flip

#flip Allele1 and Allele2 if Freq1 is bigger than .5 and change Freq1 to negative
stat_sum = statsum.copy()
cond = (stat_sum['Freq1'] > 0.5)
stat_sum.loc[cond, ['Allele1', 'Allele2']] = stat_sum.loc[cond, ['Allele2', 'Allele1']].values
#stat_sum_flip.Allele1, stat_sum_flip.Allele2 = np.where(stat_sum_flip.Freq1 > 0.5, [stat_sum_flip.Allele2, stat_sum_flip.Allele1], [stat_sum_flip.Allele1, stat_sum_flip.Allele2])

#change Freq1
stat_sum.loc[cond, ['Freq1']] = 1-stat_sum['Freq1']

#change Effect value
stat_sum.loc[cond, 'Effect'] *= -1
##no need to change  StdErr, P-value

#drop two rows Freq1 is smaller than 0.01
stat_sum.drop(stat_sum[stat_sum.Freq1 < 0.01].index, inplace=True)
stat_sum = stat_sum.reset_index(drop=True)
#stat_sum = stat_sum.drop(['index'], axis=1)

#drop unwanted columns
stat_sum.drop(stat_sum.iloc[:,10:16], axis=1, inplace=True)
stat_sum.drop(stat_sum.iloc[:,3:7], axis=1, inplace=True)

#check the unique number of MarkerName in meta
print("Unique MarkerName in meta data:", stat_sum['MarkerName'].nunique())

stat_sum

#save stat_sum 
stat_sum.to_csv(r'./data_folder/sumstat.txt', sep='\t' ,index=False)

def shell_do(command, log=False, return_log=False):
    print(f'Executing: {(" ").join(command.split())}', file=sys.stderr)

    res=subprocess.run(command.split(), stdout=subprocess.PIPE)

    if log:
        print(res.stdout.decode('utf-8'))
    if return_log:
        return(res.stdout.decode('utf-8'))

#open bim file 
bim_df = pd.read_csv(f"{geno_path}_hg19_lifted.bim", sep='\t', header=None,  names=('chr', 'rsid', 'kb', 'pos', 'a1', 'a2'))

#combine chr and pos columns
bim_df['MarkerName'] = str('chr') + bim_df.chr.astype(str) + ':' + bim_df.pos.astype(str)

#check the shape of meta
print("shape of bim data:", bim_df.shape)

#check the shape of meta
print("number of unique MarkerName in bim data:", bim_df['MarkerName'].nunique())
print("number of unique rsid in bim data:", bim_df['rsid'].nunique())
bim_df.head()

#print duplicated rows
bim_df_dup = bim_df[bim_df['MarkerName'].duplicated(keep=False)]
print('number of unique MarkerName in duplicated bim_df: ', bim_df_dup['MarkerName'].nunique())
bim_df_dup
## there are 46 duplicated MarkerName with 2 different rsid

print('Unique allele 1 in duplicated bim_df: ', bim_df_dup['a1'].unique())
print('Unique allele 2 in duplicated bim_df: ', bim_df_dup['a2'].unique())
#duplicated marker name
dup_marker = bim_df_dup['MarkerName'].unique().tolist()

#print gwas summery stat that has duplicated MarkerName
stat_sum_dup = stat_sum[stat_sum['MarkerName'].isin(dup_marker)]
stat_sum_dup

#print bim_df_dup that has duplicated MarkerName to see rsid
double_marker  = stat_sum_dup['MarkerName'].tolist()
bim_df_dup_exist = bim_df_dup[bim_df_dup['MarkerName'].isin(double_marker)]

#sort the dataframe by MarkerName
bim_df_dup_exist = bim_df_dup_exist.sort_values(by = 'MarkerName')

bim_df_dup_exist

#merge stat_sum with bim_df
meta_bim = pd.merge(stat_sum, bim_df, on=['MarkerName'], how='inner')
meta_bim

#check duplicated rsid
meta_bim_dup = meta_bim[meta_bim['rsid'].duplicated(keep=False)]
print('number of unique rsid in duplicated meta_bim: ', meta_bim_dup['rsid'].nunique())
#check number of unique rsid in case
print('number of unique rsid in meta_bim: ', meta_bim.rsid.nunique()) 
#check number of unique MarkerName in case
print('number of unique MarkerName in meta_bim: ', meta_bim.MarkerName.nunique())

#print gwas summery stat with SNP that has duplicated MarkerName
snp_stat_dup = meta_bim[meta_bim['MarkerName'].isin(dup_marker)]
snp_stat_dup
#7 duplicated (MarkerName) position with different rsid

#drop rows in meta if rsid is NaN 
#meta_df = snp_stat[snp_stat['SNP'].notna()]

sumstat_df = meta_bim.copy()
#change Allele1 and Allele2 to A1 and A2
sumstat_df = sumstat_df.rename(columns={'Allele1':'A1', 'Allele2':'A2','rsid':'SNP'})
sumstat_df = sumstat_df.drop(['a1', 'a2', 'kb'], axis=1)

#reset index
#meta_df = meta_df.reset_index(drop=True)

#check numbers of rows in meta_df
print("numbers of rows in meta data:", sumstat_df.shape[0])

sumstat_df.head()

#calculate z score with effect size and standard error (z = (score-mean)/se)

def z_score(a, b):
    if b == 0:
        return 0.0
    return float(a)/b

sumstat_df['Z'] = np.vectorize(z_score)(sumstat_df['Effect'], sumstat_df['StdErr'])

sumstat_df

#function to move columns
def movecol(df, cols_to_move=[], ref_col='', place='After'):
    
    cols = df.columns.tolist()
    if place == 'After':
        seg1 = cols[:list(cols).index(ref_col) + 1]
        seg2 = cols_to_move
    if place == 'Before':
        seg1 = cols[:list(cols).index(ref_col)]
        seg2 = cols_to_move + [ref_col]
    
    seg1 = [i for i in seg1 if i not in seg2]
    seg3 = [i for i in cols if i not in seg1 + seg2]
    
    return(df[seg1 + seg2 + seg3])

#rearrange Gene_Symbol, Chr, Start, Stop columns right nextg to index column 
sumstat = movecol(sumstat_df, cols_to_move=['SNP'], ref_col='MarkerName', place='After')

#drop MarkerName, Effect, StdErr, P-value colunms 
sumstat = sumstat.drop(['MarkerName', 'Effect', 'StdErr', 'P-value', 'chr', 'pos'], axis=1)

sumstat = sumstat.reset_index(drop=True)
print("shape of meta data:", sumstat.shape)
sumstat.head()

##sumstat is ready to use

#calculate chisq
stat_sum_chi = sumstat_df.copy()
stat_sum_chi['chisq'] = stat_sum_chi['Z']**2
print(stat_sum_chi)

#calculate lambda (median(z^2) / 0.455)
lambda_stat_sum = ((stat_sum_chi['chisq'].median())/0.455)
print("lambda using chisq: ", lambda_stat_sum)

#calculate lambda (median(z^2) / 0.455)
lambda_statsum = ((stat_sum_chi['P-value'].median())/0.456)
print("lambda using p-value: ", lambda_statsum)

#save meta_df as txt -- run only once for download
sumstat.to_csv(r'./data_folder/meta.txt', sep='\t' ,index=False)
#gwas sumary stat is ready to use

Run this after post processing

##grab significant GWAS hits (p <= 5e-8)
sig_meta = sumstat_df[sumstat_df['P-value']< 5e-8]
#reset index
sig_meta = sig_meta.reset_index(drop=True)
print("shape of significant GWAS data:", sig_meta.shape)

sig_meta.head()

sig_meta.pos.dtype

#check if double marker are in sig_meta
sig_meta_dup = sig_meta[sig_meta['MarkerName'].isin(double_marker)]
sig_meta_dup

#save sig_meta -- run only once
sig_meta.to_csv(r'./data_folder/gwas_pos.csv', index=False)

#bring in coordinate data
coords = pd.read_csv(coord_path, sep='\t')
coords

#create pos_less and pos_more which are position plus minus 1 megabase
sig_meta['pos_less']= sig_meta['pos'] - 500000
sig_meta['pos_more']= sig_meta['pos'] + 500000
sig_meta

Read the 90 parkinson's GWAS hits from Mike's paper (some are in the supplement) ##https://www.thelancet.com/journals/laneur/article/PIIS1474-4422(19)30320-5/fulltext

#read 90 parkinson's GWAS hits
pd_hits = pd.read_csv(r"./data_folder/PD_GRS_dbSNPs.tab", sep='\t', lineterminator='\r', header=None,  names=('rsid', 'author', 'condition', 'name') )
pd_hits.drop(pd_hits.iloc[:,1:4], axis=1, inplace=True)
pd_hits['rsid'] = pd_hits['rsid'].str.replace('\n', '')
pd_hits['rsid'].replace('', np.nan, inplace=True)
pd_hits.dropna(subset=['rsid'], inplace=True)
pd_hits

#merge rsid with pos from meta data

pd_hits_pos_1 = pd.merge(bim_df, pd_hits, on=['rsid'], how='inner')
pd_hits_pos_1

#get the rsid without pos
mis_rsid = pd_hits[~pd_hits['rsid'].isin(pd_hits_pos_1['rsid'])]
mis_rsid = mis_rsid.reset_index(drop=True)
mis_rsid

#merge rsid with pos from meta data

pd_hits_pos_2 = pd_hits.merge(bim_df, on=['rsid'], how='left') 
pd_hits_pos_2 = pd_hits_pos_2.drop(['kb'], axis=1)
pd_hits_pos_2

#input data for rs114138760
pd_hits_pos_2.loc[pd_hits_pos_2.rsid == 'rs114138760', 'chr'] = '1'
pd_hits_pos_2.loc[pd_hits_pos_2.rsid == 'rs114138760', 'pos'] = '154898185'
pd_hits_pos_2.loc[pd_hits_pos_2.rsid == 'rs114138760', 'a1'] = 'C'
pd_hits_pos_2.loc[pd_hits_pos_2.rsid == 'rs114138760', 'a2'] = 'G'
pd_hits_pos_2.loc[pd_hits_pos_2.rsid == 'rs114138760', 'MarkerName'] = 'chr1:154898185'

#input data for rs6658353
pd_hits_pos_2.loc[pd_hits_pos_2.rsid == 'rs6658353', 'chr'] = '1'
pd_hits_pos_2.loc[pd_hits_pos_2.rsid == 'rs6658353', 'pos'] = '161469054'
pd_hits_pos_2.loc[pd_hits_pos_2.rsid == 'rs6658353', 'a1'] = 'C'
pd_hits_pos_2.loc[pd_hits_pos_2.rsid == 'rs6658353', 'a2'] = 'G'
pd_hits_pos_2.loc[pd_hits_pos_2.rsid == 'rs6658353', 'MarkerName'] = 'chr1:161469054'

#input data for rs76116224
pd_hits_pos_2.loc[pd_hits_pos_2.rsid == 'rs76116224', 'chr'] = '2'
pd_hits_pos_2.loc[pd_hits_pos_2.rsid == 'rs76116224', 'pos'] = '18147848'
pd_hits_pos_2.loc[pd_hits_pos_2.rsid == 'rs76116224', 'a1'] = 'T'
pd_hits_pos_2.loc[pd_hits_pos_2.rsid == 'rs76116224', 'a2'] = 'A'
pd_hits_pos_2.loc[pd_hits_pos_2.rsid == 'rs76116224', 'MarkerName'] = 'chr2:18147848'

#input data for rs2042477
pd_hits_pos_2.loc[pd_hits_pos_2.rsid == 'rs2042477', 'chr'] = '2'
pd_hits_pos_2.loc[pd_hits_pos_2.rsid == 'rs2042477', 'pos'] = '96000943'
pd_hits_pos_2.loc[pd_hits_pos_2.rsid == 'rs2042477', 'a1'] = 'T'
pd_hits_pos_2.loc[pd_hits_pos_2.rsid == 'rs2042477', 'a2'] = 'A'
pd_hits_pos_2.loc[pd_hits_pos_2.rsid == 'rs2042477', 'MarkerName'] = 'chr2:96000943'

#input data for rs11683001
pd_hits_pos_2.loc[pd_hits_pos_2.rsid == 'rs11683001', 'chr'] = '2'
pd_hits_pos_2.loc[pd_hits_pos_2.rsid == 'rs11683001', 'pos'] = '102396963'
pd_hits_pos_2.loc[pd_hits_pos_2.rsid == 'rs11683001', 'a1'] = 'A'
pd_hits_pos_2.loc[pd_hits_pos_2.rsid == 'rs11683001', 'a2'] = 'T'
pd_hits_pos_2.loc[pd_hits_pos_2.rsid == 'rs11683001', 'MarkerName'] = 'chr2:102396963'

#input data for rs11707416
pd_hits_pos_2.loc[pd_hits_pos_2.rsid == 'rs11707416', 'chr'] = '3'
pd_hits_pos_2.loc[pd_hits_pos_2.rsid == 'rs11707416', 'pos'] = '151108965'
pd_hits_pos_2.loc[pd_hits_pos_2.rsid == 'rs11707416', 'a1'] = 'A'
pd_hits_pos_2.loc[pd_hits_pos_2.rsid == 'rs11707416', 'a2'] = 'T'
pd_hits_pos_2.loc[pd_hits_pos_2.rsid == 'rs11707416', 'MarkerName'] = 'chr3:151108965'

#input data for rs34025766
pd_hits_pos_2.loc[pd_hits_pos_2.rsid == 'rs34025766', 'chr'] = '4'
pd_hits_pos_2.loc[pd_hits_pos_2.rsid == 'rs34025766', 'pos'] = '17968811'
pd_hits_pos_2.loc[pd_hits_pos_2.rsid == 'rs34025766', 'a1'] = 'A'
pd_hits_pos_2.loc[pd_hits_pos_2.rsid == 'rs34025766', 'a2'] = 'T'
pd_hits_pos_2.loc[pd_hits_pos_2.rsid == 'rs34025766', 'MarkerName'] = 'chr4:17968811'

#input data for rs6825004
pd_hits_pos_2.loc[pd_hits_pos_2.rsid == 'rs6825004', 'chr'] = '4'
pd_hits_pos_2.loc[pd_hits_pos_2.rsid == 'rs6825004', 'pos'] = '77110365'
pd_hits_pos_2.loc[pd_hits_pos_2.rsid == 'rs6825004', 'a1'] = 'G'
pd_hits_pos_2.loc[pd_hits_pos_2.rsid == 'rs6825004', 'a2'] = 'C'
pd_hits_pos_2.loc[pd_hits_pos_2.rsid == 'rs6825004', 'MarkerName'] = 'chr4:77110365'

#input data for rs26431
pd_hits_pos_2.loc[pd_hits_pos_2.rsid == 'rs26431', 'chr'] = '5'
pd_hits_pos_2.loc[pd_hits_pos_2.rsid == 'rs26431', 'pos'] = '102365794'
pd_hits_pos_2.loc[pd_hits_pos_2.rsid == 'rs26431', 'a1'] = 'C'
pd_hits_pos_2.loc[pd_hits_pos_2.rsid == 'rs26431', 'a2'] = 'G'
pd_hits_pos_2.loc[pd_hits_pos_2.rsid == 'rs26431', 'MarkerName'] = 'chr5:102365794'

#input data for rs76949143
pd_hits_pos_2.loc[pd_hits_pos_2.rsid == 'rs76949143', 'chr'] = '7'
pd_hits_pos_2.loc[pd_hits_pos_2.rsid == 'rs76949143', 'pos'] = '66009851'
pd_hits_pos_2.loc[pd_hits_pos_2.rsid == 'rs76949143', 'a1'] = 'A'
pd_hits_pos_2.loc[pd_hits_pos_2.rsid == 'rs76949143', 'a2'] = 'T'
pd_hits_pos_2.loc[pd_hits_pos_2.rsid == 'rs76949143', 'MarkerName'] = 'chr7:66009851'

#input data for rs6476434
pd_hits_pos_2.loc[pd_hits_pos_2.rsid == 'rs6476434', 'chr'] = '9'
pd_hits_pos_2.loc[pd_hits_pos_2.rsid == 'rs6476434', 'pos'] = '34046391'
pd_hits_pos_2.loc[pd_hits_pos_2.rsid == 'rs6476434', 'a1'] = 'T'
pd_hits_pos_2.loc[pd_hits_pos_2.rsid == 'rs6476434', 'a2'] = 'C'
pd_hits_pos_2.loc[pd_hits_pos_2.rsid == 'rs6476434', 'MarkerName'] = 'chr9:34046391'

#input data for rs6497339
pd_hits_pos_2.loc[pd_hits_pos_2.rsid == 'rs6497339', 'chr'] = '16'
pd_hits_pos_2.loc[pd_hits_pos_2.rsid == 'rs6497339', 'pos'] = '19277493'
pd_hits_pos_2.loc[pd_hits_pos_2.rsid == 'rs6497339', 'a1'] = 'T'
pd_hits_pos_2.loc[pd_hits_pos_2.rsid == 'rs6497339', 'a2'] = 'A'
pd_hits_pos_2.loc[pd_hits_pos_2.rsid == 'rs6497339', 'MarkerName'] = 'chr16:19277493'

#input data for rs2904880
pd_hits_pos_2.loc[pd_hits_pos_2.rsid == 'rs2904880', 'chr'] = '16'
pd_hits_pos_2.loc[pd_hits_pos_2.rsid == 'rs2904880', 'pos'] = '28944396'
pd_hits_pos_2.loc[pd_hits_pos_2.rsid == 'rs2904880', 'a1'] = 'G'
pd_hits_pos_2.loc[pd_hits_pos_2.rsid == 'rs2904880', 'a2'] = 'C'
pd_hits_pos_2.loc[pd_hits_pos_2.rsid == 'rs2904880', 'MarkerName'] = 'chr16:28944396'

#input data for rs666463
pd_hits_pos_2.loc[pd_hits_pos_2.rsid == 'rs666463', 'chr'] = '17'
pd_hits_pos_2.loc[pd_hits_pos_2.rsid == 'rs666463', 'pos'] = '76425480'
pd_hits_pos_2.loc[pd_hits_pos_2.rsid == 'rs666463', 'a1'] = 'T'
pd_hits_pos_2.loc[pd_hits_pos_2.rsid == 'rs666463', 'a2'] = 'A'
pd_hits_pos_2.loc[pd_hits_pos_2.rsid == 'rs666463', 'MarkerName'] = 'chr17:76425480'

#convert column values to int
col_case=pd_hits_pos_2.columns[1:3]
pd_hits_pos_2.loc[:,col_case]=pd_hits_pos_2.loc[:,col_case].astype(np.int64)

pd_hits_pos_2

#save pd_hits_pos_2 
pd_hits_pos_2.to_csv(r'./data_folder/PD_GRS_dbSNPs_position.csv', index=False)

#create pos_less and pos_more which are position plus minus 1 megabase
pd_hits_pos_2['pos_less']= pd_hits_pos_2['pos'] - 1000000
pd_hits_pos_2['pos_more']= pd_hits_pos_2['pos'] + 1000000
pd_hits_pos_2
#read 90 parkinson's GWAS hits
#pd_hits_pos_2 = pd.read_csv(r"./data_folder/PD_GRS_dbSNPs_position.csv")
#pd_hits_pos_2

#grab gene ID from coords file if position is within start and end in coords ---> coords['start'] <= row['pos'] <= coords['end']

from numpy import nan

def match_labels(row):
    curr_df = coords[(coords['X.Chr'] == row['chr']) & ((coords['start'] <= row['pos']) & (coords['end'] >= row['pos']))]
    try:
        row['ID'] = curr_df['ID'].iloc[0]
    except:
        row['ID'] = np.NaN

    return row

result_1 = pd_hits_pos_2.apply(lambda x:match_labels(x),axis=1)

result_1

#grab gene ID from coords file if position is within start and end in coords ---> coords['start']<= row['pos_more'] <= coords['end'] or coords['end'] >= row['pos_less'] >= coords['start']

from numpy import nan

def match_labels(row):
    curr_df = coords[(coords['X.Chr'] == row['chr']) & (((coords['start'] <= row['pos_more']) & (coords['end']) >= row['pos_more']) | ((coords['start'] <= row['pos_less']) & (coords['end'] >= row['pos_less'])))]
    try:
        row['ID'] = curr_df['ID'].iloc[0]
    except:
        row['ID'] = np.NaN

    return row

result_2 = pd_hits_pos_2.apply(lambda x:match_labels(x),axis=1)

result_2

#grab gene ID from coords file if position is within start and end in coords ---> row['pos_less'] <= coords['start'] <= row['pos_more'] or row['pos_less'] <= coords['end'] <= row['pos_more']

from numpy import nan

def match_labels(row):
    curr_df = coords[(coords['X.Chr'] == row['chr']) & (((coords['start'] <= row['pos_more']) & (coords['start'] >= row['pos_less'])) | ((coords['end'] >= row['pos_less']) & (coords['end'] <= row['pos_more'])))]
    try:
        row['ID'] = curr_df['ID'].iloc[0]
    except:
        row['ID'] = np.NaN

    return row

result_3 = pd_hits_pos_2.apply(lambda x:match_labels(x),axis=1)

result_3

###row['pos_less'] <= coords['start'] <= row['pos_more'] or row['pos_less'] <= coords['end'] <= row['pos_more']

print('Null ID in reult_1: ',result_1.ID.isnull().sum())

#get non null values
print('Non Null ID in reult_1: ',result_1[['ID']].T.stack().tolist())
#result_1[['ID']].T.stack().values

print('Null ID in reult_2: ',result_2.ID.isnull().sum())

#get non null values
print('Non Null ID in reult_2: ',result_2[['ID']].T.stack().tolist())
#result_1[['ID']].T.stack().values

print('Null ID in reult_3: ',result_3.ID.isnull().sum())

#get non null values
print('Non Null ID in reult_3: ',result_3[['ID']].T.stack().tolist())
#result_1[['ID']].T.stack().values

#drop all columns except ID
gwas_genes_1 = result_1.copy()
gwas_genes_1.drop(gwas_genes_1.iloc[:,0:8], axis=1, inplace=True)
gwas_genes_1.dropna(inplace=True)
gwas_genes_1 = gwas_genes_1.reset_index(drop=True)

gwas_genes_2 = result_2.copy()
gwas_genes_2.drop(gwas_genes_2.iloc[:,0:8], axis=1, inplace=True)
gwas_genes_2.dropna(inplace=True)
gwas_genes_2 = gwas_genes_2.reset_index(drop=True)

gwas_genes_3 = result_3.copy()
gwas_genes_3.drop(gwas_genes_3.iloc[:,0:8], axis=1, inplace=True)

#check common IDs between result_1~3
gene_1 = gwas_genes_1.ID.tolist()
gene_2 = gwas_genes_2.ID.tolist()
gene_3 = result_3.ID.tolist()

print('number of ID in result_1: ', len(gene_1))
print('number of ID in result_2: ', len(gene_2))
print('number of ID in result_3: ', len(gene_3))

def comprehension(a, b):
    return [x for x in a if x in b]

list_1 = comprehension(gene_1,gene_2)
list_2 = comprehension(gene_1,gene_3)
list_3 = comprehension(gene_2,gene_3)

print('number of ID in result_1 and result_2: ', len(list_1))
print('number of ID in result_1 and result_3: ', len(list_2))
print('number of ID in result_2 and result_3: ', len(list_3))
#since all of gene_2 are in gene_3 we will combine gene_1 and gene_3

#merge gwas_genes_1 and gwas_genes_3
gwas = pd.merge(gwas_genes_1, gwas_genes_3, on='ID', how='outer')
gwas = gwas.drop_duplicates(['ID']).reset_index(drop=True)
gwas.shape

#save gwas
gwas.to_csv(r'./data_folder/PD_genes.csv', index=False)
